Fix #225 - Move all W009 checks into TripDescriptorValidator
* For optimization purposes, some of W009 checks for stop_time_updates were implemented in StopTimeUpdateValidator.  However, this currently causes problems when the *Validator classes are executed in BackgroundTask, as each class inserts a set of occurrences for the rules into the database, which caused a duplicate entry for occurrences of W009 (instead of all W009 occurrences properly being inserted in a single entry).  Moving all checks for the same rule into the same class is the workaround solution implemented in this patch.  In the future we should examine allowing more than one *Validator class to check for the same rule, which would require modifications to BackgroundTask to combine occurrence lists from multiple *Validator classes before they are inserted into the database.
