+++ src/main/java/edu/usf/cutr/gtfsrtvalidator/validation/rules/StopTimeUpdateValidator.java
@@ -24,7 +24,6 @@ import edu.usf.cutr.gtfsrtvalidator.api.model.OccurrenceModel;
 import edu.usf.cutr.gtfsrtvalidator.background.GtfsMetadata;
 import edu.usf.cutr.gtfsrtvalidator.helper.ErrorListHelperModel;
 import edu.usf.cutr.gtfsrtvalidator.util.GtfsUtils;
-import edu.usf.cutr.gtfsrtvalidator.util.RuleUtils;
 import edu.usf.cutr.gtfsrtvalidator.validation.interfaces.FeedEntityValidator;
 import org.onebusaway.gtfs.impl.GtfsDaoImpl;
 import org.slf4j.LoggerFactory;
@@ -32,15 +31,12 @@ import org.slf4j.LoggerFactory;
 import java.util.ArrayList;
 import java.util.List;
 
-import static edu.usf.cutr.gtfsrtvalidator.util.GtfsUtils.getStopTimeUpdateId;
-import static edu.usf.cutr.gtfsrtvalidator.util.GtfsUtils.getTripId;
 import static edu.usf.cutr.gtfsrtvalidator.validation.ValidationRules.*;
 
 /**
  * E002 - stop_time_updates for a given trip_id must be sorted by increasing stop_sequence
  * E036 - Sequential stop_time_updates have the same stop_sequence
  * E037 - Sequential stop_time_updates have the same stop_id
- * W009 - schedule_relationship not populated (for StopTimeUpdate)
  */
 public class StopTimeUpdateValidator implements FeedEntityValidator {
 
@@ -52,7 +48,6 @@ public class StopTimeUpdateValidator implements FeedEntityValidator {
         List<OccurrenceModel> e002List = new ArrayList<>();
         List<OccurrenceModel> e036List = new ArrayList<>();
         List<OccurrenceModel> e037List = new ArrayList<>();
-        List<OccurrenceModel> w009List = new ArrayList<>();
 
         for (GtfsRealtime.FeedEntity entity : entityList) {
             if (entity.hasTripUpdate()) {
@@ -62,7 +57,6 @@ public class StopTimeUpdateValidator implements FeedEntityValidator {
                 List<Integer> stopSequenceList = new ArrayList<>();
                 Integer previousStopSequence = null;
                 String previousStopId = null;
-                boolean foundW009 = false;
                 for (GtfsRealtime.TripUpdate.StopTimeUpdate stopTimeUpdate : stopTimeUpdateList) {
                     if (previousStopSequence != null) {
                         checkE036(entity, previousStopSequence, stopTimeUpdate, e036List);
@@ -74,13 +68,6 @@ public class StopTimeUpdateValidator implements FeedEntityValidator {
                     previousStopId = stopTimeUpdate.getStopId();
                     if (stopTimeUpdate.hasStopSequence()) {
                         stopSequenceList.add(stopTimeUpdate.getStopSequence());
-                    }
-                    // Only flag one occurrence of W009 for stop_time_update per trip to avoid flooding the database
-                    if (!foundW009) {
-                        checkW009(entity, stopTimeUpdate, w009List);
-                        if (!w009List.isEmpty()) {
-                            foundW009 = true;
-                        }
                     }
                 }
 
@@ -105,9 +92,6 @@ public class StopTimeUpdateValidator implements FeedEntityValidator {
         }
         if (!e037List.isEmpty()) {
             errors.add(new ErrorListHelperModel(new MessageLogModel(E037), e037List));
-        }
-        if (!w009List.isEmpty()) {
-            errors.add(new ErrorListHelperModel(new MessageLogModel(W009), w009List));
         }
         return errors;
     }
@@ -155,20 +139,6 @@ public class StopTimeUpdateValidator implements FeedEntityValidator {
             OccurrenceModel om = new OccurrenceModel(prefix.toString());
             errors.add(om);
             _log.debug(om.getPrefix() + " " + E036.getOccurrenceSuffix());
-        }
-    }
-
-    /**
-     * Checks rule W009 - "schedule_relationship not populated", and adds any warnings that are found to the provided warning list
-     *
-     * @param entity         entity which contains the specified trip.stop_time_update
-     * @param stopTimeUpdate stop_time_update to examine to see if it has a schedule_relationship
-     * @param warnings       list to add any warnings for W009 to
-     */
-    private void checkW009(GtfsRealtime.FeedEntity entity, GtfsRealtime.TripUpdate.StopTimeUpdate stopTimeUpdate, List<OccurrenceModel> warnings) {
-        if (stopTimeUpdate != null && !stopTimeUpdate.hasScheduleRelationship()) {
-            // W009 - schedule_relationship not populated
-            RuleUtils.addW009Occurrence(getTripId(entity, entity.getTripUpdate().getTrip()) + " " + getStopTimeUpdateId(stopTimeUpdate) + " (and potentially more for this trip)", warnings);
         }
     }
 }
